fine I guess

+++ source/AI.cpp
@@ -2580,7 +2580,7 @@ void AI::DoSurveillance(Ship &ship, Command &command, shared_ptr<Ship> &target) 
 				}
 				// Hacky way of differentiating ship behaviour without additional storage,
 				// while keeping it consistent for each ship.
-				long seed = reinterpret_cast<long>(&ship);
+				intptr_t seed = reinterpret_cast<intptr_t>(&ship);
 				int behaviour = abs(seed % 23);
 				Angle delta = Angle(360. / (behaviour / 2. + 2) * (behaviour % 2 ? -1 : 1));
 				Angle target = Angle(ship.Position()) + delta;
